requirements.txt
django-python3-ldap==0.11.3
django-ses==0.8.14
dj-database-url==0.5.0
kiwitcms-github-app==1.0
kiwitcms-tenants==1.1.1
raven==6.10.0
social-auth-app-django==3.1.0
social-auth-kerberos==0.2.4
